+++ ChangeLog.md
@@ -11,6 +11,7 @@
     - collor 100%
     - deltatime 98%
     - UnitConverter 100%
+    - config manager 100%
 
 - Data Prosesing
     - csv data pipeline 100%